requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
google-generativeai
pypdf
python-dotenv
langchain-community
langchain-text-splitters
langchain-community 
faiss-cpu 
sentence-transformers
